Refactor 'tweet_manager' and 'search_tweets'
'tweet_manager' was implementing the tweet extraction using several
buffers and a callback that collects the buffer when it grows up to a
certain size. This implementation was complex and difficult to reason
about. This commit changes the implementation to a generator that
returns one tweet at a time, which simplifies a lot the API and the code
itself.

Some changes in 'search_tweets' were necessary to adapt the new API.

Calling the configuration arguments 'args' can be misleading as 'args'
is a common Python term. This commit renames that to 'config' to make it
more clear.

Improve formatting using black (the code formatter)